+++ spec/hamli/parser_spec.rb
@@ -485,5 +485,19 @@ RSpec.describe Hamli::Parser do
         )
       end
     end
+
+    context 'with non-escaped output line' do
+      let(:source) do
+        <<~HAML
+          != a
+        HAML
+      end
+
+      it 'returns expected S-expression' do
+        is_expected.to eq(
+          [:multi, [:hamli, :position, 3, 4, [:hamli, :output, false, 'a', [:multi, [:newline]]]]]
+        )
+      end
+    end
   end
 end